backend/requirements.txt
attrs==23.2.0
blinker==1.8.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Flask==3.0.3
Flask-Cors==4.0.1
flask-expects-json==1.7.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
referencing==0.35.1
rpds-py==0.18.1
SQLAlchemy==2.0.29
tokenizer==3.4.3
typing_extensions==4.12.1
Werkzeug==3.0.2
